64bit plugins tests
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2806 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ ViewerController.m
@@ -12309,7 +12309,7 @@ int i,j,l;
 
 		//--------------------------Background color---------------------------------
 		if( [[printSettings cellWithTag: 3] state]) [settings setObject: @"YES" forKey: @"backgroundColor"];
-		float r, g, b;		
+		CGFloat r, g, b;		
 		NSColor	*rgbColor = [[printColor color] colorUsingColorSpaceName: NSDeviceRGBColorSpace];		
 		[rgbColor getRed:&r green:&g blue:&b alpha:0L];
 		[settings setObject: [NSNumber numberWithFloat: r] forKey: @"backgroundColorR"];
